+++ client-angular/src/app/core/services/platform.service.ts
@@ -9,7 +9,7 @@ import { isPlatformBrowser, isPlatformServer } from '@angular/common';
   providedIn: 'root',
 })
 export class PlatformService {
-  constructor(@Inject(PLATFORM_ID) private platformId: Object) {}
+  constructor(@Inject(PLATFORM_ID) private platformId: object) {}
 
   /**
    * Check if the code is running in a browser